# CHANGE LOG 0.2.1 - MAJOR RELEASE
Major refactoring of many systems. Items are now (almost) fully referenced by the InventoryItem table, and not the items table. Needed to be done for item modding.

Speaking of, Weapon modding is here! Fully functional weapon modding system, weapon stat calculations based on mods. Bugs are bound to be prominent, but this is the first iteration of this system.

Items now have weight that properly apply to the user when it's in their inventory.

Database Session Manager, thank you Guilherme Caminha for this article:

https://praciano.com.br/fastapi-and-async-sqlalchemy-20-with-pytest-done-right.html

Smaller improvements all around the codebase.

### Testing Improvements

- Improved some tests, nothing significantly new.

## TODO

### Testing

- [x] Top-Level routes
- [x] Auth Routes
- [x] Crew
- [x] Raids
- [x] Item/Inventory
- [x] Market
- [ ] New Weapon System

### Game Systems
- [x] Randomly generated "Rooms"
- [ ] Expanded loot pools
- [~] Combat system
- [~] Weapon Attachment system
- [ ] NPC Quests
- [ ] NPC Barter System
- [?] Scav System